Time: 195 ms (92.65%), Space: 99.1 MB (48.84%) - LeetHub

+++ 2013-detect-squares/2013-detect-squares.cpp
@@ -1,6 +1,6 @@
 class DetectSquares {
 public:
-    unordered_map<int, unordered_map<int, int>> umap;
+    unordered_map<short, unordered_map<short, short>> umap;
     DetectSquares() {
         ios_base::sync_with_stdio(false);
         cin.tie(NULL);
